Move sketch to an appropriately named folder
The Arduino IDE requires that a sketch be located in a folder of the same name. When opening a file that does not meet this requirement the Arduino IDE presents a dialog:

The file "WiFiUartTransparentBridge.ino" needs to be inside a sketch folder named "WiFiUartTransparentBridge".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE moves only the file WiFiUartTransparentBridge.ino to the new folder, leaving behind config.h. This causes compilation of the sketch to fail "fatal error: config.h: No such file or directory". As demonstrated by https://github.com/NaoNaoMe/ESP8266-WiFi-UART-transparent-bridge/issues/1, this is a genuine problem for the user.

The alternate solution of changing the sketch name to match the repository name is inferior because it is common for people to use GitHub's "Clone or Download > Download ZIP" feature to download sketches, rather than doing a Git clone and GitHub appends the branch/tag/commit to the downloaded folder name, which will cause the folder to still not match the sketch name.